src/test/java/dev/dkikolski/lc/solutions/P0001TwoSumSolutionTest.java
package dev.dkikolski.lc.solutions;


import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

import static dev.dkikolski.lc.solutions.P0001TwoSumSolution.twoSum;
import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.params.provider.Arguments.arguments;

class P0001TwoSumSolutionTest {
    public static Stream<Arguments> testcases() {
        return Stream.of(
                arguments(new int[]{2, 7, 11, 15}, 9, new int[]{0, 1}),
                arguments(new int[]{3, 2, 4}, 6, new int[]{1, 2}),
                arguments(new int[]{3, 3}, 6, new int[]{0, 1})
        );
    }

    @ParameterizedTest
    @MethodSource("testcases")
    void verify(final int[] input, final int target, final int[] expected) {
        assertThat(twoSum(input, target)).containsExactly(expected);
    }
}